Plugins: add sample_rate
