+++ plugins/sample_rate.yapsy-plugin
@@ -1,0 +1,9 @@
+
+[Core]
+Name = sample_rate
+Module = sample_rate
+
+[Documentation]
+Author = Various
+Version = 0.1
+Description = Print average sample rate.
